# backend/nest/apps/api-gateway/src/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RWA API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/v1"
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login with email/password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/2fa/verify": {
      "post": {
        "summary": "Verify 2FA code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFaVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA verified"
          }
        }
      }
    },
    "/kyc/submit": {
      "post": {
        "summary": "Submit KYC data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KycSubmission"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "KYC submitted"
          }
        }
      }
    },
...
    "/orderbook/{assetId}": {
      "get": {
        "summary": "Get orderbook for asset",
        "parameters": [
          {
            "in": "path",
            "name": "assetId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "side",
            "schema": {
              "type": "string",
              "enum": [
                "buy",
                "sell"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrders"
                }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Post an order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List my orders",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "side",
            "schema": {
              "type": "string",
              "enum": [
                "buy",
                "sell"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "partially_filled",
                "filled",
                "cancelled"
              ]
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrders"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "delete": {
        "summary": "Cancel order",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled"
          }
        }
      }
    },
...
    "/distributions/{assetId}": {
      "get": {
        "summary": "List distributions for asset",
        "parameters": [
          {
            "in": "path",
            "name": "assetId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Distributions"
          }
        }
      }
    },
    "/admin/distributions/trigger": {
      "post": {
        "summary": "Trigger a payout batch (admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Batch queued"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SignupRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          },
          "phone": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "TwoFaVerifyRequest": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
      "KycSubmission": {
        "type": "object",
        "properties": {
          "documentType": {
            "type": "string"
          },
          "documentNumber": {
            "type": "string"
          },
          "country": {
            "type": "string"
          }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "required": [
          "access_token"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          }
        }
      },
...
      "PaginatedOrders": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
      "OrderRequest": {
        "type": "object",
        "required": [
          "asset_id",
          "side",
          "qty",
          "price"
        ],
        "properties": {
          "asset_id": {
            "type": "integer"
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ]
          },
          "qty": {
            "type": "number"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "asset_id": {
            "type": "integer"
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ]
          },
          "qty": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "partially_filled",
              "filled",
              "cancelled"
            ]
          },
          "filled_qty": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TriggerRequest": {
        "type": "object",
        "required": [
          "asset_id",
          "period",
          "gross"
        ],
        "properties": {
          "asset_id": {
            "type": "string"
          },
          "period": {
            "type": "string",
            "description": "ISO8601 period, e.g., 2024-01/2024-03"
          },
          "gross": {
            "type": "number"
          },
          "mgmt_fee_bps": {
            "type": "integer"
          },
          "carry_bps": {
            "type": "integer"
          }
        }
      }
    }
  }
}